backend/requirements.txt
flask
flask_cors
flask_pymongo
python-dotenv
